Aggiunto Timer per la creazione della maschera di ImmagineCliccabile
Aggiunto script per la rimozione degli IP che non rispondono dalla lista server

+++ QRisiko/Gestore Servers.cpp
@@ -11,6 +11,7 @@ GestoreServers::GestoreServers(QObject *parent)
 
 void GestoreServers::getFile(const QUrl &url)
 {
+	errore=false;
 	file.setFileName("TempServerList.txt");
 	http.setHost(url.host(), url.port(80));
 	http.get(url.path(), &file);
@@ -18,9 +19,14 @@ void GestoreServers::getFile(const QUrl &url)
 }
 void GestoreServers::httpDone(bool error)
 {
-	if (error) QMessageBox::critical(0,tr("Errore"),tr("Errore durante la comunicazione con il server"));
-	file.close();
-	if (ToDelete) file.remove();
+	if (error) {
+		QMessageBox::critical(0,tr("Errore"),tr("Errore durante la comunicazione con il server"));
+		errore=true;
+	}
+	else{
+		file.close();
+		if (ToDelete) file.remove();
+	}
 	emit done();
 }
 void GestoreServers::AddIP(){
@@ -38,16 +44,18 @@ void GestoreServers::OttieniLista(){
 }
 
 void GestoreServers::FormServerList(){
-	QString ServerList("");
-	if(file.open(QIODevice::ReadOnly)) {
-		QTextStream in(&file);
-		while(!in.atEnd()) {
-			ServerList.append(in.readLine());        	
+	if(!errore){
+		QString ServerList("");
+		if(file.open(QIODevice::ReadOnly)) {
+			QTextStream in(&file);
+			while(!in.atEnd()) {
+				ServerList.append(in.readLine());        	
+			}
 		}
+		file.close();
+		emit ListaOttenuta(ServerList);
+		file.remove();
 	}
-	file.close();
-	emit ListaOttenuta(ServerList);
-	file.remove();
 	disconnect(this,SIGNAL(done()),this,SLOT(FormServerList()));
 }
 void GestoreServers::OttieniIP(){
@@ -56,15 +64,23 @@ void GestoreServers::OttieniIP(){
 	connect (this,SIGNAL(done()),this,SLOT(FormIP()));
 }
 void GestoreServers::FormIP(){
-	QString ServerList("");
-	if(file.open(QIODevice::ReadOnly)) {
-		QTextStream in(&file);
-		while(!in.atEnd()) {
-			ServerList.append(in.readLine());        	
+	if (!errore)
+	{
+		QString ServerList("");
+		if(file.open(QIODevice::ReadOnly)) {
+			QTextStream in(&file);
+			while(!in.atEnd()) {
+				ServerList.append(in.readLine());        	
+			}
 		}
+		file.close();
+		emit IPOttenuto(ServerList);
+		file.remove();
 	}
-	file.close();
-	emit IPOttenuto(ServerList);
-	file.remove();
 	disconnect(this,SIGNAL(done()),this,SLOT(FormIP()));
+}
+
+void GestoreServers::NotResponding(QString ipnr){
+	ToDelete=true;
+	getFile(QUrl(Giocatori::PathToHttpServer+"/NotResponding.php?IPNotResponding="+ipnr));
 }
